Addresses issue with non-numerical sub- and super-scripts.
This patch allows us to use a broader range of Unicode symbols with sub-scripts and super-scripts. I was having issues with Greek letters in sub-script in one of my projects and this is fixing the problem. See the example below:

julia> latexify(:(2a₍ₐᵦᵧ₎+ΓhΛ₁₂- ħᵗⁱᵦ))
L"$2 \cdot a{_{( a \beta \gamma )}} + {\Gamma}h\Lambda{_{1 2}} - \hbar{^{t i}}{_\beta}$"

Many thanks for this great library btw!